# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "L5 OpenApi",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "darius@matulionis.lt"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0",
        "x-logo": {
            "url": "https://via.placeholder.com/190x90.png?text=L5-Swagger"
        }
    },
    "paths": {
        "/api/accounts": {
            "get": {
                "tags": [
                    "accounts"
                ],
                "summary": "Returns all accounts",
                "description": "Returns a map of status codes to quantities",
                "operationId": "getAccounts",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "isSuccessful": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "hasContent": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Account"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "isSuccessful": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "hasContent": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "401"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "unauthenticated"
                                        },
                                        "data": {
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "accounts"
                ],
                "summary": "Stores one account",
                "description": "Returns a map of status codes to quantities",
                "operationId": "storeAccount",
                "requestBody": {
                    "description": "Pass account credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@test.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "PassWord12345"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "422": {
                        "description": "Wrong credentials response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "isSuccessful": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "hasContent": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "422"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "invalid credentials"
                                        },
                                        "detailed_error": {
                                            "example": "null"
                                        },
                                        "data": {
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "isSuccessful": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "hasContent": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Account"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "isSuccessful": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "hasContent": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "401"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "unauthenticated"
                                        },
                                        "data": {
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/api/accounts/1": {
            "get": {
                "tags": [
                    "accounts"
                ],
                "summary": "Returns one account by id",
                "description": "Returns a map of status codes to quantities",
                "operationId": "getAccount",
                "responses": {
                    "200": {
                        "description": "added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "isSuccessful": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "hasContent": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "200"
                                        },
                                        "message": {
                                            "type": "string"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Account"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "isSuccessful": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "hasContent": {
                                            "type": "boolean",
                                            "example": "false"
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": "401"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "unauthenticated"
                                        },
                                        "data": {
                                            "example": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Account": {
                "title": "Account model",
                "description": "Account model",
                "properties": {
                    "email": {
                        "title": "email",
                        "type": "string",
                        "default": "admin@test.com"
                    },
                    "phone_number": {
                        "title": "phone_number",
                        "type": "string",
                        "default": "0966666666"
                    },
                    "password": {
                        "title": "password",
                        "type": "string",
                        "default": "123456"
                    }
                },
                "type": "object"
            }
        }
    },
    "security": [
        []
    ]
}